requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=/mnt/c/workspace/brewhaus/requirements.txt /mnt/c/workspace/brewhaus/requirements.in
#
alembic==1.5.8
    # via -r /mnt/c/workspace/brewhaus/requirements.in
click==7.1.2
    # via
    #   -r /mnt/c/workspace/brewhaus/requirements.in
    #   flask
flask-login==0.5.0
    # via -r /mnt/c/workspace/brewhaus/requirements.in
flask==1.1.2
    # via
    #   -r /mnt/c/workspace/brewhaus/requirements.in
    #   flask-login
greenlet==1.0.0
    # via sqlalchemy
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
mako==1.1.4
    # via alembic
markupsafe==1.1.1
    # via
    #   jinja2
    #   mako
mysql-connector-python==8.0.24
    # via -r /mnt/c/workspace/brewhaus/requirements.in
mysql-connector==2.2.9
    # via -r /mnt/c/workspace/brewhaus/requirements.in
protobuf==3.15.8
    # via mysql-connector-python
python-dateutil==2.8.1
    # via alembic
python-dotenv==0.17.1
    # via -r /mnt/c/workspace/brewhaus/requirements.in
python-editor==1.0.4
    # via alembic
six==1.15.0
    # via
    #   protobuf
    #   python-dateutil
sqlalchemy==1.4.12
    # via
    #   -r /mnt/c/workspace/brewhaus/requirements.in
    #   alembic
werkzeug==1.0.1
    # via flask
